requirements.txt
aniso8601==10.0.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
filelock==3.13.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
fsspec==2024.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
pillow==11.0.0
pycparser==2.22
pyOpenSSL==25.0.0
pytz==2025.1
setuptools==70.2.0
six==1.17.0
sympy==1.13.1
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchvision==0.21.0+cu118
typing_extensions==4.12.2
Werkzeug==3.1.3
